Abstracted new DbKvs tests to also be used by Oracle
